Move FrameCodec to common module for shared use
- Relocated FrameCodec from broker.infra.network to common.protocol
- Changed class visibility from package-private to public
- Moved FrameCodecTest to common module test directory
- Enables sensor-node and control-panel modules to use FrameCodec

Resolves Issue #59
